backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
httpx
psycopg[binary]
python-dotenv